coding structural changes and outage repo
